+++ src/objectwizards/symbolwizards/alSymbolsBasedPageWizard.ts
@@ -129,7 +129,7 @@ export class ALSymbolsBasedPageWizard extends ALSymbolsBasedWizard {
         tableSymbol.collectChildSymbols(AZSymbolKind.Field, true, fieldList);
         fieldList.forEach(
             item => {
-                writer.writePageField(item.name, item.name, undefined, false);
+                writer.writePageField(item.name, item.name, undefined, undefined, false);
             }
         );