+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 3.0.8
+ - Issue #256 - Suggestion: Tooltip generation from field's description
+
+Thank you
+ - mjmatthiesen for reporting issue #256
+
 ## 3.0.7
  - Issue #245 - Error during Build : al-lang-proxy\tempalfile.al(1,10) after using "go to definition"
  - Issue #252 - sortVariables / sortProperties with // comments results in invalid objects